+++ programs/bpf_loader/src/syscalls/mod.rs
@@ -511,7 +511,7 @@ fn translate(
     len: u64,
 ) -> Result<u64, Error> {
     memory_mapping
-        .map(access_type, vm_addr, len)
+        .map(access_type, vm_addr, len as usize)
         .map_err(|err| err.into())
         .into()
 }
@@ -524,7 +524,7 @@ fn translate_type_inner<'a, T>(
 ) -> Result<&'a mut T, Error> {
     let host_addr = translate(memory_mapping, access_type, vm_addr, size_of::<T>() as u64)?;
     if !check_aligned {
-        Ok(unsafe { std::mem::transmute::<u64, &mut T>(host_addr) })
+        Ok(unsafe { std::mem::transmute::<usize, &mut T>(host_addr as usize) })
     } else if !address_is_aligned::<T>(host_addr) {
         Err(SyscallError::UnalignedPointer.into())
     } else {
@@ -2431,7 +2431,7 @@ mod tests {
             0,
             &mut memory_mapping,
         );
-        assert_access_violation!(result, 0x100000001, string.len() as u64);
+        assert_access_violation!(result, 0x100000001, string.len());
         let result = SyscallLog::rust(
             &mut invoke_context,
             0x100000000,
@@ -2441,7 +2441,7 @@ mod tests {
             0,
             &mut memory_mapping,
         );
-        assert_access_violation!(result, 0x100000000, string.len() as u64 * 2);
+        assert_access_violation!(result, 0x100000000, string.len() * 2);
 
         let result = SyscallLog::rust(
             &mut invoke_context,
@@ -2748,7 +2748,7 @@ mod tests {
             0,
             &mut memory_mapping,
         );
-        assert_access_violation!(result, rw_va - 1, HASH_BYTES as u64);
+        assert_access_violation!(result, rw_va - 1, HASH_BYTES);
         let result = SyscallHash::rust::<Sha256Hasher>(
             &mut invoke_context,
             ro_va,
